Calculating popular tweets
I have calculated the number of likes and retweets for a tweet, and the total number of likes for all retweets. By 'all', I mean every level of retweets, so if a tweet is retweeted multiple times, the number of likes for each retweet will be included in the calculation.